Address feedback: remove old eth_ledger_bridge_keyring, rename eth_ledger_bridge_keyring_new
